+++ src/main/java/com/o2/site/O2Application.java
@@ -1,8 +1,10 @@
-package com.o2.site.config;
+package com.o2.site;
 
+import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+@MapperScan
 @SpringBootApplication
 public class O2Application {
