+++ src/main/java/com/o2/site/member/main/controller/HomeController.java
@@ -1,0 +1,23 @@
+package com.o2.site.member.main.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+    @GetMapping("/login")
+    public String login() {
+        return "member/login";
+    }
+
+    @GetMapping("/join")
+    public String join() {
+        return "member/join";
+    }
+
+    @GetMapping("/myPage")
+    public String myPage() {
+        return "member/myPage";
+    }
+}
