--- src/main/java/com/o2/site/member/main/dao/UserDAO.java
@@ -1,4 +1,0 @@
-package com.o2.site.member.main.dao;
-
-public class UserDAO {
-}
