--- src/main/java/com/o2/site/member/main/domain/UserRepository.java
@@ -1,9 +1,0 @@
-package com.o2.site.member.main.domain;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class UserRepository {
-
-
-}
