--- src/main/java/com/o2/site/member/main/service/UserService.java
@@ -1,12 +1,0 @@
-package com.o2.site.member.main.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-
-//    private final UserRepository userRepository;
-
-
-
-}
