--- src/main/java/com/o2/site/member/main/service/main/dao/UserDAO.java
@@ -1,4 +1,0 @@
-package com.o2.site.member.main.service.main.dao;
-
-public class UserDAO {
-}
