--- src/main/java/com/o2/site/member/main/service/main/domain/User.java
@@ -1,18 +1,0 @@
-package com.o2.site.member.main.service.main.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long M;
-
-    private String
-
-
-}
