+++ src/main/resources/application.yaml
@@ -9,7 +9,11 @@ spring:
     username: scott
     password: tiger
 
-# mybatis 설정
+# mybatis 설정 (SQL 쿼리의 결과 매핑을 자동으로 처리)
 mybatis:
   configuration:
     map-underscore-to-camel-case: true
+      # true 이면 마이바티스는 데이터베이스 컬럼 이름을 자바 객채의 Camel Case 필드와 매핑할 때 언더스코어('_')를 무시하고
+      # Camel Case로 변환한다.
+      # ex) 데이터베이스 컬럼 이름이 'first_name'인 경우 자바 객체의 필드명은 'firstName'으로 매핑됨
+  mapper-locations: classpath:mybatis/mappers/*.xml
