+++ src/main/resources/static/css/login/loginpage.css
@@ -61,3 +61,6 @@
 .form__field:required, .form__field:invalid {
     box-shadow: none;
 }
+
+
+
