--- src/main/resources/templates/layout/joinLayout.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html th:fragment="layout (title, content, insertHead)" xmlns:th="http://www.thymeleaf.org"
-      xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity6">
-<head>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css">
-    <title th:replace="${title}">타이틀</title>
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js"
-            integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r"
-            crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js"
-            integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy"
-            crossorigin="anonymous"></script>
-    <th:block th:replace="${insertHead}"></th:block>
-    <style>
-        /* 푸터 스타일링 */
-        footer {
-            margin-top: 100px;
-            background-color: #343a40;
-            color: #fff;
-            padding: 20px 0;
-            width: 100%;
-        }
-
-        .footer-content {
-            text-align: center;
-        }
-
-        .footer-content p {
-            margin-bottom: 0;
-        }
-    </style>
-</head>
-<body>
-<header class="d-flex flex-wrap align-items-center justify-content-center justify-content-md-between py-3 mb-4 border-bottom px-3">
-    <a href="/" class="d-flex align-items-center col-md-3 mb-2 mb-md-0 text-dark text-decoration-none">
-        O2
-    </a>
-
-    <ul class="nav col-12 col-md-auto mb-2 justify-content-center mb-md-0">
-        <li><a href="/" class="nav-link px-2 link-secondary">홈</a></li>
-        <li><a href="/" class="nav-link px-2 link-dark">페이지1</a></li>
-        <li><a href="/" class="nav-link px-2 link-dark">페이지2</a></li>
-        <li><a href="/" class="nav-link px-2 link-dark">페이지3</a></li>
-    </ul>
-
-    <div class="col-md-3 text-end">
-        <a href="/login" type="button" class="btn btn-outline-primary me-2">로그인</a>
-        <a href="/join" type="button" class="btn btn-primary">회원가입</a>
-        <a href="/login" type="button" class="btn btn-outline-primary me-2">로그아웃</a>
-    </div>
-</header>
-<main th:replace="${content}">
-    템플릿메인
-</main>
-<footer>
-    <div class="container footer-content">
-        <p>o2마켓 &copy; 2024. All rights reserved.</p>
-        <p>세미 프로젝트</p>
-    </div>
-</footer>
-</body>
-</html>
