+++ src/main/resources/templates/member/join.html
@@ -1,11 +1,9 @@
 <!DOCTYPE html>
-<html th:replace="~{layout/joinLayout :: layout(~{::title}, ~{::main}, ~{::insertHead})}"
+<html th:replace="~{layout/mainLayout :: layout(~{::title}, ~{::main}, ~{::insertHead})}"
       xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
     <title>회원가입</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
-          integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
     <th:block th:fragment="insertHead">
         <link rel="stylesheet" type="text/css" href="../static/css/login/loginpage.css" th:href="@{/css/login/loginpage.css}">
         <link rel="stylesheet" type="text/css" href="../static/css/login/joinpage.css" th:href="@{/css/login/joinpage.css}">
@@ -13,25 +11,23 @@
 </head>
 <body>
 <main>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe"
-            crossorigin="anonymous"></script>
-    <div class="login-container">
+    <div class="join-container">
         <form action="/join" method="post">
-            <h1>환영합니다.</h1>
+            <h1>회원가입.</h1>
             <div class="form__group field">
                 <label for="userId" class="form__label">아이디</label>
                 <input type="text" id="userId" name="id" class="form__field" required>
             </div>
             <div class="form__group field">
                 <!-- id로 css, 자바스크립트 연결 가능-->
+                <!-- name 속성은 이후 자바 객체와 매핑될 이름 -->
                 <label for="userPw" class="form__label">비밀번호</label>
                 <input type="password" id="userPw" name="password" class="form__field" required>
             </div>
             <div class="form__group field">
                 <!-- id로 css, 자바스크립트 연결 가능-->
                 <label for="userPwCheck" class="form__label">비밀번호 확인</label>
-                <input type="password" id="userPwCheck" name="password_Check" class="form__field" required>
+                <input type="password" id="userPwCheck" name="passwordCheck" class="form__field" required>
             </div>
             <div class="form__group field">
                 <label for="userName" class="form__label">이름</label>
@@ -39,7 +35,7 @@
             </div>
             <div class="form__group field">
                 <label for="userPhone" class="form__label">전화번호</label>
-                <input type="text" id="userPhone" name="Phone_Number" class="form__field" required>
+                <input type="text" id="userPhone" name="phoneNumber" class="form__field" required>
             </div>
             <div class="form__group field">
                 <label for="userAddress" class="form__label">주소</label>
